[build] Strip sqlite DSO on Release builds
Fixes: https://github.com/xamarin/xamarin-android/issues/4244

Turns out our `Release` build failed to strip the `sqlite3` shared
library (it was built with `-g`) and so it ended up packaged with all
the debug symbols included, thus unnecessarily growing the size of APKs
we produce.

Pass `-s` to the compiler during `Release` builds so that only the
dynamic symbols are present in the resulting shared library and no debug
symbols are present.

This decreases the shared library around 5 times.