[Xamarin.Android.Build.Tasks] Update multidex warnings & make localizable
Context: 0342fe5698b86e21e36c924732ff135b9a87e4af
Context: https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1009374/

[3757be20][0] moved the XA4305 messages for `<CompileToDalvik/>` into
the `.resx` file, but it overlooked the uses of XA4305 in
`<CreateMultiDexMainDexClassList/>` and `<R8/>`.

The original uses of XA4305 in `<CreateMultiDexMainDexClassList/>` and
`<R8/>` were a little confusing because the `@(MultiDexMainDexList)`
MSBuild item corresponds to the `CustomMainDexListFiles` property in the
tasks, not the `MultiDexMainDexListFile` property.

To reduce the potential for confusion between `@(MultiDexMainDexList)`
and `MultiDexMainDexListFile` and to separate them in telemetry, give
the warning message related to `@(MultiDexMainDexList)` its own new
XA4309 warning code, reword the other XA4305 messages, and update the
documentation accordingly.

Along the way, move the messages for the new XA4309 code and the other
remaining multidex message code XA4306 into the `.resx` file so that
they are localizable.

[0]: https://github.com/xamarin/xamarin-android/commit/3757be20ba9fd1f34166f812daf9963f5d6dd22b