[xabuild] Make xaprepare actually ignore installed Mono version
`xaprepare` has two flags to ignore maximum and minimum mono versions:

     --ignore-max-mono-version
     --ignore-min-mono-version

which, up until now, required that you pass a boolean value to them as a
parameter which is both unnecessary and annoying. This commit makes the
value optional and flag's presence on command line without the specified
value makes it default to `true`.

Also added is a flag, `--ignore-mono-version` which sets **both** of the
above flags to, again, optional boolean value passed to this argument.

The `PREPARE_IGNORE_MONO_VERSION` variable in top-level Makefile hasn't
been actually used in the build. Make it pass `--ignore-mono-version` to
`xaprepare` if set to a value other than `0`.