[Xamarin.Android.Build.Tasks] mono-symbolicate warnings are messages (#4258)
Fixes: https://github.com/xamarin/xamarin-android/issues/2654

If you build a Xamarin.Forms template in `Release` with
`MonoSymbolArchive=True`, you get many warnings such as:

	Warning: Directory obj\Release\android\assets contains Xamarin.Android.Arch.Core.Common.dll but no debug symbols file was found.
	Warning: Directory obj\Release\android\assets contains Xamarin.Android.Arch.Core.Runtime.dll but no debug symbols file was found.
	...
	Warning: Directory obj\Release\android\assets contains Xamarin.Forms.Platform.dll but no debug symbols file was found.

These warnings surface as MSBuild warnings with no error code.  These
are coming from NuGet packages and there is not a way for developers
to "fix" the warning.  It is not actionable.

I took the time to clean up how we invoke `mono-symbolicate`:

	<Exec
	    Command="&quot;$(MonoAndroidBinDirectory)mono-symbolicate&quot; store-symbols &quot;$(_MSYMDirectory)&quot; &quot;$(IntermediateOutputPath)android/assets&quot;"
	    Condition=" '$(MonoSymbolArchive)' == 'True' "
	/>

I made a `<MonoSymbolicate />` MSBuild task that uses a `Regex` to
downgrade the specific MSBuild warning to a message.

I updated a test to call `mono-symbolicate` to verify we don't get
warnings going forward.