[build] Consolidate Mono dependency information
We have a bit of a disconnect with respect to our dependency on Mono
across different tools. `xaprepare` has some Mono version information
embedded in it, and we duplicate some of this information in
`Configuration.props`. Furthermore, we don't do an adequate job of
enforcing a version ceiling in our Azure Pipelines builds.

For additional context, see https://github.com/xamarin/xamarin-android/pull/4219
for all of the places that require updating when we bump the version of
Mono that we depend on.

In order to improve this story, `xaprepare` has been updated to gather
all Mono dependency information from `Configuration.props`, and no
longer specifies any version information of its own. We will also now
roll back the installation of Mono on CI builds if the installed version
exceeds the max version specified in this file. The following properties
in `Configuration.props` should be updated when bumping the version of
Mono that we build against:

    `<MonoDarwinPackageUrl>` - The package to install if the version of Mono on disk is not within the expected range (as defined below).
    `<MonoRequiredMinimumVersion>` - The lowest version of Mono that we will build against.
    `<MonoRequiredMaximumVersion>` - The highest version of Mono that we will build against.