[build] Consolidate Mono dependency information (#4282)
Context: https://github.com/xamarin/xamarin-android/pull/4219

We have a bit of a disconnect with respect to our dependency on Mono
across different tools.  `xaprepare` has some Mono version information
embedded in it, and we duplicate some of this information within
`Configuration.props`.  Furthermore, we don't do an adequate job of
enforcing a version ceiling in our Azure Pipelines builds, resulting
in some PR builds against master using mono/2020-02 from PR #4219,
not the expected mono/2019-12 master *should* be using.

For additional context, see PR #4219 for all of the places that
require updating when we bump the version of Mono that we depend on.

In order to improve this story, `xaprepare` has been updated to gather
all Mono dependency information from `Configuration.props`, and no
longer specifies any version information of its own. We will also now
roll back the installation of Mono on CI builds if the installed
version exceeds the maximum version specified in this file.  The
following properties in `Configuration.props` should be updated when
bumping the version of Mono that we build against:

  * `$(MonoDarwinPackageUrl)`: The macOS `.pkg` file to install if the
    version of the "system Mono" is not within the expected range,
    as defined below.

  * `$(MonoRequiredMinimumVersion)`: The lowest supported "system mono"
    version that we will build against.

  * `$(MonoRequiredMaximumVersion)`: The highest supported version of
    "system mono" that we will build against.