[CI] Ensure Mono integration branch PRs run full regression suite (#4295)
In commit 3e0b4da9 we created a new YAML variable called `IsMonoBranch`
that checks if the branch name contains `mono-` so we could determine
if it's a mono integration branch.  If it is, we would like to run the
full regression test suite on it at the PR stage and not wait until the
branch is merged.

This is because a Mono build is riskier and has a higher chance of
breaking stuff.

However, the way this was implemented only works for manually run
builds (which is how I tested it and confirmed that it works):

	contains(variables['Build.SourceBranchName'], 'mono-')

If the build is kicked off automatically from the PR,
[`Build.SourceBranchName` is overwritten][0] with the value
`"merge"`.  Thus we have not been running these suites on the in-
progress Mono 2020-02 integration branch (PR #4219).

In order to fix this, also check the variable
`System.PullRequest.SourceBranch`, which will contain the branch name
when this is a PR initiated build.

[0]: https://github.com/microsoft/azure-pipelines-tasks/issues/8793