[docs] Edit new BuildProcess items for formatting and wording
Indent link reference definitions in list items.  This is not required,
but it makes it easier to confirm by eye that all the list items follow
the expected style.  And CommonMark does allow link reference
definitions to occur inside of block containers like list items.

Wrap the lines for the `@(AndroidResourceAnalysisConfig)` entry, and add
an "Added in" sentence for it.

Adjust "Added in" sentences to Xamarin.Android 10.2 to account for
cherry-picks to d16-5 from master for:
`StartAndroidActivity`, `StopAndroidPackage`,
`BuildAndStartAotProfiling`, `FinishAotProfiling`,
`$(AndroidExtraAotOptions)`, `$(AndroidDebugStoreType)`

Move the "Added in" sentences to the end of the list items in cases
where they weren't already.

Replace the term "the launch activity" because that term [doesn't appear
on developer.android.com][0].  "The default activity" is a little
imprecise, but it does appear on developer.android.com, and it has
roughly the correct meaning.

Adjust wording for clarity and grammar.

Add information from release notes where applicable.

[Replace "e.g."][1]

[0]: https://www.google.com/search?q="the+launch+activity"+site%3Adeveloper.android.com
[1]: https://docs.microsoft.com/style-guide/a-z-word-list-term-collections/e/eg