[Xamarin.Android.Build.Tasks] Re-enable Resources.Designer.cs gen (#4222)
Now that `Xamarin.Android.Build.Tasks.csproj` is a short-form project
-- see commit 03743f32 -- some item metadata needs to be added
explicitly to `Resources.resx` and `Resources.Designer.cs` to tell
Visual Studio to regenerate `Resources.Designer.cs` whenever
`Resources.resx` changes, as expected.  The short-form project style
does add the files automatically to the appropriate
`@(EmbeddedResource)` and `@(Compile)` item lists via the default
wildcard includes, but it doesn't add this metadata.

The new items are almost the same as in commit 0342fe56, but they use
the `Update` attribute instead of the `Include` attribute because now
the items are already included via the wildcard includes.