Bump $(ProductVersion) to 10.2.100
Commit aa8eba4b was a mistake, accidentally cherry-picked from
xamarin/xamarin-android/master@584369a5.

This was later caught and reverted in 23d9cb41, but that still leaves
a problem with versioning within Visual Studio for Windows:
Version Numbers Cannot Go Backwards.

Unfortunately, the `revision` value is computed based on the number of
commits since `$(ProductVersion)` was last changed, and since 23d9cb41
changed `$(ProductVersion)`, that means it will get v10.2.99.0, which
is considerably *less than* the last installer inserted into VS.

There are thus two paths forward:

 1. Take this as an opportunity to bump to the "real" version of
    v10.3.0.

 2. Use a different version smaller <= v10.3.0 but >= v10.2.99.

Preview 2 is supposed to be Feature Complete™, so I personally like
the idea of d16-6 Preview 2 being v10.3.0.0 (or thereabouts).
As such, go with (2) and set `$(ProductVersion)` to v10.2.100.