[Xamarin.Android.Build.Tests] Add SmokeTests category (#4269)
Add a new `SmokeTests` category to `Xamarin.Android.Build.Tests.dll`
as part of a broader effort to help split up our test execution and
improve turnaround times for our Azure Pipelines builds.  My selection
of tests to add to this category was somewhat arbitrary, but it
includes a handful of build, rebuild, and incremental build scenarios,
as well as multiple project configurations (including AOT and Bundle
Assemblies), and native/java build actions.

We've also been seeing a somewhat rare case where the VS instance on
our Windows machines become unlaunchable and the `MSBuild@1` YAML task
falls back to a very old version, breaking the build.  I've attempted
to fix this by updating our script that reverts the version of XA
installed with VS to first try to resume or repair the Visual Studio
installation on disk.