[Xamarin.Android.Build.Tasks] Remove KeyTool from signing (#4308)
Invoking `keytool -list` takes about 200ms (my machine, best case,
likely mostly due to java startup).  @jonathanpeppers and I were
looking for ways to speed up Apply Changes builds and Android builds
in general.

The `keytool -list` invocation isn't *needed* at all.  It was added
as a convenience to address [Bug #13154][0], to make it easier to
determine the SHA1 fingerprint of the keystore used to sign the `.apk`.
It wasn't a "need", it was a "nice to have", but at the cost of 200ms
to execute, it is no longer "nice to have"; it is instead a nuisance.

Additionally:

  * KeyTool is only run when `$(AndroidKeyStore)` isn't supplied, so
    it's only run when using the default debug keystore, not when the
    user supplies one
  * I don't actually see anything output in VS (at least on Windows)
    in the build log when `keytool` is run there; the output isn't
    captured.
  * `keytool` output isn't used as other otherwise by our build
    (except for the occasional error blowing up the build)
  * If the user supplies an invalid keystore or key alias password,
    the sign task itself will fail (I did a quick test to verify).

It seems like KeyTool doesn't serve any useful purpose here & we'd
like to get the perf bump of removing it.

If you still need to determine the keystore signature, please see:

	https://docs.microsoft.com/en-us/xamarin/android/deploy-test/signing/keystore-signature

[0]: https://xamarin.github.io/bugzilla-archives/13/13154/bug.html