[Xamarin.Android.Build.Tasks] remove $(AndroidExplicitCrunch)
`$(AndroidExplicitCrunch)` is an experimental feature that was aimed
at improving incremental build times for projects containing lots of
images. Unfortunately, I have found that it also can cause projects to
run `aapt` on every build. It may (or may not) help build times
depending on the project.

I think now is a good time to remove `$(AndroidExplicitCrunch)`, since
is not supported with `aapt2`. We may be in an `aapt2`-only world soon.

Removing `$(AndroidExplicitCrunch)` will make it easier for us to
refactor things going forward. Projects using it will continue to
work, but just use the default `aapt` behavior.

I also removed some tests with:

    [Ignore ("Enable once Merge Resources is finished")]

This was a "merge resources" feature that was never finished. It
relied on `$(AndroidExplicitCrunch)`, so we can remove these tests.