[Xamarin.Android.Build.Tests] Move debug.keystore into Project Tree
When running unit tests we occasionally see a random
`ANDKT0000` build error. We suspect this happens when
two tests try to create the same `debug.keystore` in
the same locaiton.

To work around this issue all projects should use override
the `_ApkDebugKeyStore` property to make sure the file
is created in the project directory. This should stop
parallel tests creating the same file at the same time.