[build] Preserve CONFIGURATION for subsequent make invocations
When the tree is configured with `make CONFIGURATION=Release`, every
subsequent `make` invocation will require `CONFIGURATION=Release` to be
present on the command line or otherwise the build tree will be messed
up and will require a re-run of `make prepare`.
Prevent the above from happening by storing the configuration in a
`bin/configuration.mk` file which is included by the `Makefile` as the
very first thing.