[ci] Always set artifact run number, even if previous steps failed (#4404)
In commit af2aa5b4 we append the job run number to the log artifact's
name to create a unique name so that the upload succeeds.

However the script does not have `condition: always()` set so it
isn't run if a previous step failed.  This results in similar
failures when uploading the `xa-build-status*` and `xa-test-results*`
log files as what af2aa5b4 attempted to fix.

Always run the "Set upload artifact name" step.