[Xamarin.Android.Build.Tasks] don't use temp files in <ConvertResourcesCases/> (#4395)
I noticed that we were still using temporary files for
`AndroidResource.UpdateXmlResource()`.  I reworked this to use
`MemoryStreamPool.CreateStreamWriter()` and generate the file in-memory.

We also had the code:

	if (srcmodifiedDate <= lastUpdate) {
	    Log.LogDebugMessage ("  Skipping: {0}  {1} <= {2}", file, srcmodifiedDate, lastUpdate);
	    continue;
	}
	Log.LogDebugMessage ("  Processing: {0}   {1} > {2}", file, srcmodifiedDate, lastUpdate);

I removed the "Skipping" log message, because we have found that
MSBuild log messages can hurt performance.  In this case, it could
potentially log a lot of files, and it will help incremental builds
to only log files that changed.

Building the Toggl app on Windows:

https://github.com/jonathanpeppers/mobileapp/tree/peppers-fixes

Originally from:

https://github.com/toggl/mobileapp

  * Before:

        613 ms  ConvertResourcesCases                      2 calls

  * After:

        508 ms  ConvertResourcesCases                      2 calls

For this app, this will save ~100ms on an initial build, or builds
where Android resources change.

I also cleaned up a few things:

  * Removed an unused `logMessage` parameter.
  * Removed `ConvertResourcesCases.AdditionalResourceDirectories` so
    that both aapt & aapt2 will invoke `<ConvertResourcesCases/>` the
    same way.