[Xamarin.Android.Build.Tests] Move debug.keystore into Project Tree (#4381)
Sometimes, the `CodeBehindUnitTests.BuildTests_CodeBehindBuildTests`
fail with an ANDKT0000 error:

	Task "AndroidCreateDebugKey"
	  Task Parameter:KeyStore=…/.local/share/Xamarin/Mono for Android/debug.keystore
	  Task Parameter:KeyAlias=androiddebugkey
	  Task Parameter:KeyPass=android
	  Task Parameter:StorePass=android
	  Task Parameter:StoreType=pkcs12
	  Task Parameter:KeyAlgorithm=RSA
	  Task Parameter:Validity=10950
	  Task Parameter:ToolPath=…/bin
	  Task Parameter:Command=-genkeypair
	  KeyTool : -genkeypair
	            …/.local/share/Xamarin/Mono for Android/debug.keystore
	  …/bin/keytool -genkeypair -alias androiddebugkey -storepass android -keypass android -keystore "…/.local/share/Xamarin/Mono for Android/debug.keystore" -dname "CN=Android Debug,O=Android,C=US" -keyalg RSA -validity 10950 -storetype pkcs12
	…Xamarin.Android.Common.targets(2877,2): error ANDKT0000: keytool error: java.io.EOFException 
	  The command exited with code 1.

The current theory is that the `CodeBehindUnitTests` test fixture is
running the tests potentially in parallel, which may cause them to
all attempt to create the *same* `debug.keystore` file at the "same"
time, which could explain why `keytool -genkeypair` throws an
`EOFException`: the file is being written to concurrently.

Attempt to work around this issue by having all test projects override
the `$(_ApkDebugKeyStore)` property to ensure that the `debug.keystore`
file is created in the project directory.  This should prevent the
parallel tests from creating the same file at the same time.