[docs] Add missing items to messages README
Add some missing items to the error and warning messages README so that
the list includes all the numbers that are currently in
`src/Xamarin.Android.Build.Tasks/Properties/Resources.resx` as well as
all of the numbers that have existing files under
`Documentation/guides/messages/`.

Other changes:

Add a documentation file for XA2002.

Add the recently added XA4217 Java.Interop error code to the README.