[Mono.Android] Dont include Test code in Mono.Android.dll (#4433)
* [Mono.Android] Dont include Test code in Mono.Android.dll

I recently hit the following test failure

	CheckTimestamps:
	obj/Release/android/assets/it-IT/Mono.Android.resources.dll` is older than `20/03/2020 11:43:38`, with a timestamp of `20/03/2020 11:04:48`!

Which was completely weird. Partly because the project was not
building any localised assets. But mostly because it was
for `Mono.Android`.

It turns out the move to SDK style projects was accidently
including ALL the `.cs` files from the `Test` folder as
well as the normal code.

This exaplins why we are now generating this file and including
it in our build system.

So lets Remove all the `Test` files from the project.

* Exclude None and EmbeddedResource too