[Xamarin.Android.Sdk] fixes for $(PackageVersion)
I found with ae136f0f, was that I was getting the following issue on
Windows:

    "C:\src\xamarin-android\src\Xamarin.Android.Sdk\Xamarin.Android.Sdk.Lite.proj" (Pack target) (1) ->
        Xamarin.Android.Sdk\Xamarin.Android.Sdk.Lite.proj(5,9): error MSB4184:
        The expression "[System.IO.File]::ReadAllText(..\..\bin\Debug\Version)" cannot be evaluated. Could not find file 'C:\bin\Debug\Version'.

It seems the property:

    <PackageVersion Condition="Exists('..\..\bin\$(Configuration)\Version')">$([System.IO.File]::ReadAllText('..\..\bin\$(Configuration)\Version'))</PackageVersion>

Somehow the evaluation of `Exists` and `ReadAllText` are doing
different things? Weird.

I refactored things to use `$(MSBuildThisFileDirectory)` and
intermediate properties for the file paths.

After that... I found that the value of `$(PackageVersion)` had line
breaks, such as:

    PackageVersion = 10.3.99
    .126

^^ included a trailing line break as well.

I added a `.Trim()` call for both `File.ReadAllText` calls.